SEARCH-263 norsk-magic & internasjonal-magic ParallelSearchCommandExecutor uses Executors.newCachedPool Enrichments now keep and sort on a float analysis score value.
git-svn-id: http://sesat.no/svn/sesat-kernel/trunk@2855 cdc3010f-a60b-0410-91f3-d90f869209c8

+++ src/java/no/schibstedsok/front/searchportal/query/run/RunningQueryImpl.java
@@ -269,8 +269,8 @@ public class RunningQueryImpl extends AbstractRunningQuery implements RunningQue
                         final SearchConfiguration config = searchResult.getSearchCommand().getSearchConfiguration();
                         final String name = config.getName();
                         final SearchTab.EnrichmentHint eHint = context.getSearchTab().getEnrichmentByCommand(name);
-                        final int score = scores.get(name) != null
-                                ? (int)(scores.get(name) * eHint.getWeight())
+                        final float score = scores.get(name) != null
+                                ? scores.get(name) * eHint.getWeight()
                                 : 0;
 
                         // update hit status
